refactor: 각 뷰에 맞춰 패키지 변경
