+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MovieReservationActivity"
+            android:name=".moviereservation.MovieReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".ReservationResultActivity"
+            android:name=".reservationresult.ReservationResultActivity"
             android:exported="false" />
         <activity
-            android:name=".ScreeningMovieActivity"
+            android:name=".screeningmovie.ScreeningMovieActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
