+++ app/src/main/java/woowacourse/movie/reservationresult/ReservationResultActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.reservationresult
 
 import android.content.Context
 import android.content.Intent
@@ -6,10 +6,8 @@ import android.os.Bundle
 import android.view.MenuItem
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.R
 import woowacourse.movie.data.StubMovieRepository
-import woowacourse.movie.model.ReservationResultUiModel
-import woowacourse.movie.presenter.ReservationResultPresenter
-import woowacourse.movie.view.ReservationResultView
 
 class ReservationResultActivity : AppCompatActivity(), ReservationResultView {
     private lateinit var presenter: ReservationResultPresenter
