+++ app/src/main/java/woowacourse/movie/reservationresult/ReservationResultView.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.reservationresult
+
+import woowacourse.movie.reservationresult.ReservationResultUiModel
+
+interface ReservationResultView {
+    fun showResult(reservationResult: ReservationResultUiModel)
+}
