+++ app/src/main/java/woowacourse/movie/screeningmovie/ScreenMovieUiModel.kt
@@ -1,13 +1,12 @@
-package woowacourse.movie.model
+package woowacourse.movie.screeningmovie
 
 import androidx.annotation.DrawableRes
 import woowacourse.movie.R
 
-data class MovieReservationUiModel(
+data class ScreenMovieUiModel(
     val id: Long,
     val title: String,
     @DrawableRes val imageRes: Int = R.drawable.img_movie_poster,
     val screenDate: String,
-    val description: String,
     val runningTime: String,
 )
