+++ app/src/main/res/layout/activity_detail_movie.xml
@@ -5,7 +5,7 @@
     android:id="@+id/detail_movie"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieReservationActivity">
+    tools:context=".moviereservation.MovieReservationActivity">
 
     <ScrollView
         android:layout_width="match_parent"
