+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScreeningMovieActivity">
+    tools:context=".screeningmovie.ScreeningMovieActivity">
 
 
     <ListView
