+++ app/src/main/res/layout/item_movie.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="12dp"
-    tools:context=".ScreeningMovieActivity">
+    tools:context=".screeningmovie.ScreeningMovieActivity">
 
     <ImageView
         android:id="@+id/iv_movie_post"
